Move code depending on `codegen-full` to its own file
